% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.nnormal.R
\name{gts.normal}
\alias{gts.normal}
\title{Apply the generalized Bayesian two-stage normal-based causal model with instrumental variables.}
\usage{
gts.normal(formula, data, advanced = FALSE, adv.model, b0 = 1,
  B0 = 1e-06, g0 = 0, G0 = 1e-06, e0 = 0.001, E0 = 0.001,
  beta.start = NULL, gamma.start = NULL, e.start = NULL,
  n.chains = 1, n.burnin = floor(n.iter/2), n.iter = 10000,
  n.thin = 1, DIC, debug = FALSE, codaPkg = FALSE)
}
\arguments{
\item{formula}{An object of class formula: a symbolic description of the model to be fitted.
The details of the model specification are given under "Details".}

\item{data}{A dataframe with the variables to be used in the model.}

\item{advanced}{Logical; if FALSE (default), the model is specified using the formula argument,
if TRUE, self-defined models can be specified using the adv.model argument.}

\item{adv.model}{Specify the self-defined model. Used when advanced=TRUE.}

\item{b0}{The mean hyperparameter of the normal distribution (prior distribution)
for the first-stage generalized causal model coefficients, i.e., coefficients for the instrumental variables.
This can either be a numerical value or a vector with dimensions equal to the number of coefficients
for the instrumental variables. If this takes a numerical value, then that values will
serve as the mean hyperparameter for all of the coefficients for the instrumental variables.
Default value of 0 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{B0}{The precision hyperparameter of the normal distribution (prior distribution)
for the first stage generalized causal model coefficients.
This can either be a numerical value or a vector with dimensions equal to the number of coefficients
for the instrumental variables. If this takes a numerical value, then that values will
serve as the precision hyperparameter for all of the coefficients for the instrumental variables.
Default value of 1.0E-6 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{g0}{The mean hyperparameter of the normal distribution (prior distribution)
for the second-stage generalized causal model coefficients,
i.e., coefficients for the treatment variable and other regression covariates).
This can either be a numerical value if there is only one treatment variable in the model,
or a if there is a treatment variable and multiple regression covariates,
with dimensions equal to the total number of coefficients for the treatment variable and covariates.
Default value of 0 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{G0}{The precision hyperparameter of the normal distribution (prior distribution)
for the second-stage generalized causal model coefficients.
This can either be a numerical value if there is only one treatment variable in the model,
or a vector if there is a treatment variable and multiple regression covariates,
with dimensions equal to the total number of coefficients for the treatment variable and covariates.
Default value of 1.0E-6 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{e0}{The location hyperparameter of the inverse Gamma distribution (prior for the variance of the
normal distribution on the model residual).
Default of 0.001 is equivalent to the noninformative prior for the inverse Gamma distribution.}

\item{E0}{The shape hyperparameter of the inverse Gamma distribution (prior for the variance of the
normal distribution on the model residual).
Default of 0.001 is equivalent to the noninformative prior for the inverse Gamma distribution.}

\item{beta.start}{The starting values for the first-stage generalized causal model coefficients,
i.e., coefficients for the instrumental variables.
This can either be a numerical value or a column vector with dimensions
equal to the number of first-stage coefficients.
The default value of NA will use the IWLS (iteratively reweighted least squares) estimate
of first-stage coefficients as the starting value.
If this is a numerical value, that value will
serve as the starting value mean for all the first-stage beta coefficients.}

\item{gamma.start}{The starting values for the second-stage generalized causal model coefficients,
i.e., coefficients for the treatment variable and the model covariates.
This can either be a numerical value or a column vector with dimensions
equal to the number of second-stage coefficients.
The default value of NA will use the IWLS (iteratively reweighted least squares) estimate
of second-stage coefficients as the starting value.
If this is a numerical value, that value will
serve as the starting value mean for all the second-stage gamma coefficients.}

\item{e.start}{The starting value for the precision hyperparameter of the inverse gamma distribution
(prior for the scale parameter of Student's t distribution of the model residual).
The default value of NA will use the inverse of the residual variance from the
IWLS (iteratively reweighted least square) estimate of the second-stage model.}

\item{n.chains}{The number of Markov chains. The default is 1.}

\item{n.burnin}{Length of burn in, i.e., number of iterations to discard at the beginning.
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.iter}{The number of total iterations per chain (including burnin). The default is 10000.}

\item{n.thin}{The thinning rate. Must be a positive integer. The default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=Dbar-Dhat is used.}

\item{codaPkg}{Logical; if FALSE (default), an object is returned; if TRUE,
file names of the output are returned.}
}
\value{
If \emph{codaPkg=FALSE}(default), returns an object containing summary statistics of
the saved parameters, including
\item{s1.intercept}{Estimate of the intercept from the first stage.}
\item{s1.slopeP}{Estimate of the pth slope from the first stage. }
\item{s2.intercept}{Estimate of the intercept from the second stage.}
\item{s2.slopeP}{Estimate of the pth slope from the second stage (the first slope is always
the \strong{LATE}).}
\item{var.e.s2}{Estimate of the residual variance at the second stage.}
\item{DIC}{Deviance Information Criterion.}
If \emph{codaPkg=TRUE}, the returned value is the path for the output file
containing the Markov chain Monte Carlo output.
}
\description{
The \code{gts.nnormal} function applies the generalized Bayesian two-stage
normal-based causal model to the categorical treatment data.
The model best suits the normally-distributed outcome data
that are complete or ignorably missing (i.e., missing completely at random or missing at random).
}
\details{
\enumerate{
\item{Bayesian two-stage causal models are specified symbolically.
A typical model has the form \emph{reponse ~ terms|instrumental_variables},
where response is the (numeric) response vector and terms is a series of terms
which specifies a linear predictor (i.e., the treatment variable and the covariates) for the response.
The first specification in the term is always the treatment variable and
the remaining specifications are always the covariates for the response.}
\item{DIC is computed as \emph{mean(deviance)+pD}.}
\item{Prior distributions used in ALMOND.}
\itemize{
\item Generalized causal model coefficients at both stages: normal distributions.
\item The generalized causal model residual term: normal distribution.
}
}
}
\examples{
\donttest{
# Run the model
model1 <- gts.nnormal(outcome~treatment|instrument,data=simCatNormMCAR)

# Run the model with the self-defined advanced feature

model2 <- gts.nnormal(outcome~treatment|instrument,data=simCatNormMCAR,
advanced=TRUE, adv.model=my.model)

# Extract the model DIC
model1$DIC

# Extract the MCMC output
model3 <- gts.nnormal(outcome~treatment|instrument,data=simCatNormMCAR,codaPkg=TRUE)
}

}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Rubin, D.B. (2003).
\emph{Bayesian data analysis}, 2nd edition. Chapman and Hall/CRC Press.

Spiegelhalter, D. J., Thomas, A., Best, N. G., Gilks, W., & Lunn, D. (1996).
BUGS: Bayesian inference using Gibbs sampling.
\href{http://www.mrc-bsu.cam.ac.uk/bugs}, 19.
}
