% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts.nrobust.s.R
\name{gts.nrobust.s}
\alias{gts.nrobust.s}
\title{Apply the generalized Bayesian two-stage robust-selection causal model with instrumental variables.}
\usage{
gts.nrobust.s(formula, data, m.ind, advanced = FALSE, adv.model,
  b0 = 1, B0 = 1e-06, g0 = 0, G0 = 1e-06, e0 = 0.001,
  E0 = 0.001, l0 = 0, L0 = 1e-06, v0 = 0, V0 = 100,
  beta.start = NULL, gamma.start = NULL, e.start = NULL,
  lambda0.start = 1, lambda1.start = 1, df.start = 5, n.chains = 1,
  n.burnin = floor(n.iter/2), n.iter = 50000, n.thin = 1, DIC,
  debug = FALSE, codaPkg = FALSE)
}
\arguments{
\item{formula}{An object of class formula: a symbolic description of the model to be fitted.
The details of the model specification are given under "Details".}

\item{data}{A dataframe with the variables to be used in the model.}

\item{advanced}{Logical; if FALSE (default), the model is specified using the formula argument,
if TRUE, self-defined models can be specified using the adv.model argument.}

\item{adv.model}{Specify the self-defined model. Used when advanced=TRUE.}

\item{b0}{The mean hyperparameter of the normal distribution (prior distribution)
for the first-stage generalized causal model coefficients, i.e., coefficients for the instrumental variables.
This can either be a numerical value or a vector with dimensions equal to the number of coefficients
for the instrumental variables. If this takes a numerical value, then that values will
serve as the mean hyperparameter for all of the coefficients for the instrumental variables.
Default value of 0 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{B0}{The precision hyperparameter of the normal distribution (prior distribution)
for the first stage generalized causal model coefficients.
This can either be a numerical value or a vector with dimensions equal to the number of coefficients
for the instrumental variables. If this takes a numerical value, then that values will
serve as the precision hyperparameter for all of the coefficients for the instrumental variables.
Default value of 1.0E-6 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{g0}{The mean hyperparameter of the normal distribution (prior distribution)
for the second-stage generalized causal model coefficients,
i.e., coefficients for the treatment variable and other regression covariates.
This can either be a numerical value if there is only one treatment variable in the model,
or a if there is a treatment variable and multiple regression covariates,
with dimensions equal to the total number of coefficients for the treatment variable and covariates.
Default value of 0 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{G0}{The precision hyperparameter of the normal distribution (prior distribution)
for the second-stage generalized causal model coefficients.
This can either be a numerical value if there is only one treatment variable in the model,
or a vector if there is a treatment variable and multiple regression covariates,
with dimensions equal to the total number of coefficients for the treatment variable and covariates.
Default value of 1.0E-6 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{e0}{The location hyperparameter of the inverse Gamma distribution (prior for the scale parameter
of Student's t distribution on the model residual).
Default of 0.001 is equivalent to the noninformative prior for the inverse Gamma distribution.}

\item{E0}{The shape hyperparameter of the inverse Gamma distribution (prior for the scale parameter
of Student's t distribution on the model residual).
Default of 0.001 is equivalent to the noninformative prior for the inverse Gamma distribution.}

\item{l0}{The mean hyperparameter of the normal distribution (prior for the added-on selection model coefficients).
This can either be a numerical value or a vector with dimensions equal to the number of coefficients for the instrumental variables.
If this takes a numerical value, then that values will serve as the mean hyperparameter for all of the coefficients
for the instrumental variables. Default value of 0 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{L0}{The precision hyperparameter of the normal distribution (prior for the added-on selection model coefficients).
This can either be a numerical value or a vector with dimensions equal to the number of coefficients for the instrumental variables.
If this takes a numerical value, then that values will serve as the precision hyperparameter for all of the coefficients
for the instrumental variables.
Default value of 1.0E-6 is equivalent to a noninformative prior for the normal distributions.
Used when advanced=FALSE.}

\item{v0}{The lower boundary hyperparameter of the uniform distribution (prior for the degrees of freedom
parameter of Student's t distribution).}

\item{V0}{The upper boundary hyperparameter of the uniform distribution (prior for the degrees of freedom
parameter of Student's t distribution).}

\item{beta.start}{The starting values for the first-stage generalized causal model coefficients,
i.e., coefficients for the instrumental variables.
This can either be a numerical value or a column vector with dimensions
equal to the number of first-stage coefficients.
The default value of NA will use the IWLS (iteratively reweighted least squares) estimate
of first-stage coefficients as the starting value.
If this is a numerical value, that value will
serve as the starting value mean for all the first-stage beta coefficients.}

\item{gamma.start}{The starting values for the second-stage generalized causal model coefficients,
i.e., coefficients for the treatment variable and the model covariates.
This can either be a numerical value or a column vector with dimensions
equal to the number of second-stage coefficients.
The default value of NA will use the IWLS (iteratively reweighted least squares) estimate
of second-stage coefficients as the starting value.
If this is a numerical value, that value will
serve as the starting value mean for all the second-stage gamma coefficients.}

\item{e.start}{The starting value for the precision hyperparameter of the inverse gamma distribution
(prior for the scale parameter of Student's t distribution of the model residual).
The default value of NA will use the inverse of the residual variance from the
IWLS (iteratively reweighted least square) estimate of the second-stage model.}

\item{lambda0.start}{The starting value for the intercept of the coefficient of the added-on selection model.}

\item{lambda1.start}{The starting value for the slope of the coefficient of the added-on selection model.}

\item{df.start}{The starting value for the degrees of freedom of Student's t distribution.}

\item{n.chains}{The number of Markov chains. The default is 1.}

\item{n.burnin}{Length of burn in, i.e., number of iterations to discard at the beginning.
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.iter}{The number of total iterations per chain (including burnin). The default is 50000.}

\item{n.thin}{The thinning rate. Must be a positive integer. The default is 1.}

\item{DIC}{Logical; if TRUE (default), compute deviance, pD, and DIC. The rule pD=Dbar-Dhat is used.}

\item{codaPkg}{Logical; if FALSE (default), an object is returned; if TRUE,
file names of the output are returned.}
}
\value{
If \emph{codaPkg=FALSE}(default), returns an object containing summary statistics of
the saved parameters, including
\item{s1.intercept}{Estimate of the intercept from the first stage.}
\item{s1.slopeP}{Estimate of the pth slope from the first stage. }
\item{s2.intercept}{Estimate of the intercept from the second stage.}
\item{s2.slopeP}{Estimate of the pth slope from the second stage (the first slope is always
the \strong{LATE}).}
\item{select.intercept}{Estimate of the intercept from the added-on selection model.}
\item{select.slope}{Estimate of the slope from the added-on selection model.}
\item{var.e.s2}{Estimate of the residual variance at the second stage.}
\item{df.est}{Estimate of the degrees of freedom for the Student's t distribution.}
\item{DIC}{Deviance Information Criterion.}
If \emph{codaPkg=TRUE}, the returned value is the path for the output file
containing the Markov chain Monte Carlo output.
}
\description{
The \code{gts.nrobust.s} function applies the generalized Bayesian two-stage
robust-selection causal model to the categorical treatment data.
The model best suits the outcome data that contain outliers
and are nonignorably missing (i.e., MNAR) (e.g., dropout, attrition).
}
\details{
\enumerate{
\item{The formula takes the form \emph{response ~ terms|instrumental_variables}.}
      \code{\link{gts.nnormal}} provides a detailed description of the formula rule.
\item{DIC is computed as \emph{mean(deviance)+pD}.}
\item{Prior distributions used in ALMOND.}
\itemize{
\item Generalized causal model coefficients at both stages: normal distributions.
\item The generalized causal model residual: Student's t distribution.
\item Added-on selection model coefficients: normal distributions.
}
}
}
\examples{
\donttest{
# Run the model
model1 <- gts.nrobust.s(outcome~treatment|instrument,data=simCatOutMNAR,
m.ind=simCatOutMNAR$indicator,n.iter=100000)

# Run the model with the self-defined advanced feature
my.model <- function(){
  for (i in 1:N){
    logit(p[i]) <- beta0 + beta1*z[i]
    x[i] ~ dbern(p[i])
    muY[i] <- gamma0 + gamma1*p[i]
    y[i] ~ dt(muY[i], pre.u2, df)

    m[i] ~ dbern(q[i])
    q[i] <- phi(lambda0 + lambda1*y[i])
  }

  beta0 ~ dnorm(0,1)
  beta1 ~ dnorm(1, 1)
  gamma0 ~ dnorm(0, 1)
  gamma1 ~ dnorm(.5, 1)
  lambda0 ~ dnorm(0, 1.0E-6)
  lambda1 ~ dnorm(0, 1.0E-6)

  pre.u2 ~ dgamma(.001, .001)

  df ~ dunif(0,50)

  s1.intercept <- beta0
  s1.slope1 <- beta1
  s2.intercept <- gamma0
  s2.slope1 <- LATE
  select.intercept <- lambda0
  select.slope <- lambda1
  df.est <- df
  var.e.s2 <- 1/pre.u2
}

model2 <- ts.nrobust.s(outcome~treatment|instrument,m.ind=simCatOutMNAR$indicator,data=simCatOutMNAR,
advanced=TRUE, adv.model=my.model,n.iter=100000)

# Extract the model DIC
model1$DIC

# Extract the MCMC output
model3 <- gts.nrobust.s(outcome~treatment|instrument,data=simCatOutMNAR,
m.ind=simCatOutMNAR$indicator,n.iter=100000, codaPkg=TRUE)
}

}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Rubin, D.B. (2003).
\emph{Bayesian data analysis}, 2nd edition. Chapman and Hall/CRC Press.

Spiegelhalter, D. J., Thomas, A., Best, N. G., Gilks, W., & Lunn, D. (1996).
BUGS: Bayesian inference using Gibbs sampling.
\href{http://www.mrc-bsu.cam.ac.uk/bugs}, 19.
}
