% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simVoucher.R
\docType{data}
\name{simVoucher}
\alias{simVoucher}
\title{Simulated data of a public voucher program.}
\format{A data frame with 1954 rows (participants) and 10 columns (variables).}
\usage{
data(simVoucher)
}
\description{
The dataset is simulated from a study of the effect of a public voucher program
(Currie and Yelowitz, 2000). The dataset contains the status of the public voucher program
participation and other attributes. The variables are as follows.
}
\examples{
\donttest{
data(simVoucher)
}

}
\references{
Currie, J., & Yelowitz, A. (2000).
Are public housing projects good for kids? \emph{Journal of public economics},
\emph{75}(1), 99-124.
}
\keyword{datasets}
